requirement.txt
python-dotenv
mcp
pyodbc
loguru
openai